(BOLT-969) Exclude cache tmp directory from purge
Previously when purging the temporary direcotory for downloading task files (weekly or on startup) the `tmppath` directory itself was deleted resulting in errors downloading files. Instead of deleting the temporary directory during a purge delete only the contents of the temporary directory.

+++ lib/bolt_server/file_cache.rb
@@ -166,7 +166,7 @@ module BoltServer
       expired_time = Time.now - purge_ttl
       @cache_dir_mutex.with_write_lock do
         Dir.glob(File.join(@cache_dir, '*')).select { |f| File.directory?(f) }.each do |dir|
-          if (mtime = File.mtime(dir)) < expired_time
+          if (mtime = File.mtime(dir)) < expired_time && dir != tmppath
             @logger.debug("Removing #{dir}, last used at #{mtime}")
             FileUtils.remove_dir(dir)
           end
